Move `OsuUserInputManager` out from nested class
