+++ osu.Game/Input/OsuUserInputManager.cs
@@ -1,0 +1,38 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Input;
+using osuTK.Input;
+
+namespace osu.Game.Input
+{
+    public class OsuUserInputManager : UserInputManager
+    {
+        internal OsuUserInputManager()
+        {
+        }
+
+        protected override MouseButtonEventManager CreateButtonEventManagerFor(MouseButton button)
+        {
+            switch (button)
+            {
+                case MouseButton.Right:
+                    return new RightMouseManager(button);
+            }
+
+            return base.CreateButtonEventManagerFor(button);
+        }
+
+        private class RightMouseManager : MouseButtonEventManager
+        {
+            public RightMouseManager(MouseButton button)
+                : base(button)
+            {
+            }
+
+            public override bool EnableDrag => true; // allow right-mouse dragging for absolute scroll in scroll containers.
+            public override bool EnableClick => false;
+            public override bool ChangeFocusOnClick => false;
+        }
+    }
+}
